Step 1: Extract omnia-primitives crate
- Move Event, VectorClock, blake3_domain, WireFormat into new zero-heavy-deps crate.
- Add re-export shims in substrate for backward compatibility.
- All lib tests pass (57 primitives + 397 substrate + 58 economics + 62 shards).
- cargo check --workspace passes with zero errors.
- Primitives compiles in <6s (dev profile).
- ed25519-dalek added to primitives for Event signing/verification.
- bincode (workspace) added to primitives for legacy wire format support.
- thiserror version aligned to 2.0 to match workspace and avoid duplicate deps.
- No logic changes — only import paths and crate structure.
