+++ binding/Cargo.toml
@@ -10,6 +10,7 @@ repository = { workspace = true }
 [dependencies]
 omnia-substrate = { path = "../substrate" }
 omnia-shards = { path = "../shards" }
+omnia-primitives = { path = "../omnia-primitives" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1"
 postcard = { version = "1", features = ["alloc"] }
